feat: complete 06-build-page

+++ 04-copy-directory/index.js
@@ -1,22 +1,39 @@
 const { readdir, mkdir, copyFile } = require('fs/promises');
 const path = require('path');
 
-function copyDir() {
-  const srcPath = path.resolve(__dirname, 'files')
+function copyDir(srcPath, destPath) {
+
   readdir(srcPath, { withFileTypes: true })
     .then((files) => {
-      const destPath = path.resolve(__dirname, 'files-copy')
+
       mkdir(destPath, { recursive: true })
       for (const file of files) {
-        copyFile(
-          path.resolve(srcPath, file.name),
-          path.resolve(destPath, file.name)
-        )
+        if (file.isFile()) {
+          copyFile(
+            path.resolve(srcPath, file.name),
+            path.resolve(destPath, file.name)
+          )
+        } else {
+
+          copyDir(
+            path.resolve(srcPath, file.name),
+            path.resolve(destPath, file.name)
+          )
+        }
       }
     }, (err) => {
+
+      console.log(err)
+    })
+    .catch(err => {
+      
       console.log(err)
     })
 }
 
 
-copyDir()
+copyDir(path.resolve(__dirname, 'files'), path.resolve(__dirname, 'files-copy'))
+
+module.exports = {
+  copyDir
+}
